Add support for debugging with LLDB
[reviewed by elliot]

A new flag has been added, --lldb which works in the same manner as --gdb,
but launches lldb rather than gdb. This is primarily for anyone who is
developing on OSX (as gdb doesn't work overly well there).

Some notes for anyone who is planning on using this:

Older versions of LLDB (including the stock one on 10.9 which is
lldb-310.2.37) and chpl have some odd interactions. I'm not sure where the
fault lies, but breakpoints will not fire on the first run of the
compiler. If you run the process again in the same session the breakpoints
should work fine. This can be solved by compiling your own version of LLDB
and using that instead of the system version (I'm using lldb-320.99.0
which seems to work fine).

I've added the commands from gdb.commands to a file that gets sourced
called lldb.commands.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23501 3a8e244f-b0f2-452b-bcba-4c88e055c3ca