Remove some timeout files
[ok'd by perf team]

This is the start of a general cleaning up of the timeout files in the testing
directory. The primary reason is to get tests that actually require the timeout
down to under the default timeout to decrease testing time. 

All of these tests currently finish well under their timeouts so the timeouts
are being removed. Some of these were previously performance tests so they
might have taken longer, but overall I think compilation time/execution time
has decreased substantially for some tests in the last year or so that the
timeouts aren't needed anymore.

There's ~15 tests that still have timeouts (and another 7 that have timeouts
smaller than the default, which we are going to leave in.) 

Removed files and associated timeouts are:

  900  - arrays/deitz/parallelism/stream/TIMEOUT
  300  - npb/ep/mcahir/ep.perftimeout
  300  - npb/ft/ft-serial.perftimeout
  600  - release/examples/benchmarks/miniMD/explicit/TIMEOUT
  600  - release/examples/benchmarks/miniMD/TIMEOUT
  600  - release/examples/benchmarks/shootout/threadring.perftimeout
  600  - studies/hpcc/HPL/bradc/hpl-blc-noreindex.timeout
  600  - studies/hpcc/HPL/vass/hpl.timeout
  900  - studies/hpcc/STREAMS/bradc/stream-block1d-local.timeout
  900  - studies/hpcc/STREAMS/bradc/stream-block1d-promote.timeout
  900  - studies/hpcc/STREAMS/bradc/stream-block1d.timeout
  3200 - studies/lulesh/sungeun/TIMEOUT
  600  - studies/madness/aniruddha/madchap/test_likepy.timeout
  400  - studies/shootout/thread-ring/lydia/thread-ring-coforall-begin.perftimeout
  400  - studies/shootout/thread-ring/lydia/thread-ring-coforall.perftimeout
  400  - studies/shootout/thread-ring/lydia/thread-ring-for-begin.perftimeout
  600  - trivial/shannon/compSampler.timeout


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23503 3a8e244f-b0f2-452b-bcba-4c88e055c3ca