Re-add timeouts for threadring
This adds the timeouts that I removed from threadring yesterday. We had
timeouts for these tests on 2 of the 3 perf boxes and the last one was close to
the timeout. Rather than switch to a smaller problem size, we have decided to
put the timeouts back, and once we switch to qthreads as the default we should
be able to remove these timeouts again as qthreads is much faster for this
benchmark.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23506 3a8e244f-b0f2-452b-bcba-4c88e055c3ca