Remove some more timeout files
[ok'd by perf team]

This continues a  general cleaning up of the timeout files in the testing
directory. The primary reason is to get tests that actually require the timeout
down to under the default timeout to decrease testing time. 

All of these tests currently finish well under their timeouts so the timeouts
are being removed. Some of these were previously performance tests so they
might have taken longer, but overall I think compilation time/execution time
has decreased substantially for some tests in the last year or so that the
timeouts aren't needed anymore.

There's ~5 tests that still have timeouts (and another 7 that have timeouts
smaller than the default, which we are going to leave in.) 

Removed files and associated timeouts are:

  600  - studies/cholesky/jglewis/TIMEOUT
  600  - studies/cholesky/jglewis/version2/elemental/TIMEOUT
  600  - studies/cholesky/jglewis/version2/performance/TIMEOUT
  600  - studies/cholesky/jglewis/version2/standard_cholesky/TIMEOUT
  1800 - release/examples/benchmarks/ssca2/TIMEOUT
  1000 - studies/ssca2/jglewis/ssca2_version2/TIMEOUT
  1800 - studies/ssca2/performance/TIMEOUT
  1800 - studies/ssca2/rachels/TIMEOUT
  1800 - studies/ssca2/sungeun/TIMEOUT
  600  - distributions/dm/t5a.timeout


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23510 3a8e244f-b0f2-452b-bcba-4c88e055c3ca