Change the bounds of Space2D32 to be *not* so close to max(int(32))
[Reviewed by Elliot]

This patch changes the bounds of Space2D32 to be well within the range
of int(32) and updates .good files and one .chpl file accordingly.

Background:

A bunch of tests in the distribution robustness suite started failing
with --fast testing when we started using gcc 4.7 because their domain
bounds were butting up against max(int(32)).  FWIW, We have discussed
this issue at length and know that we have an issue with this when
iterating as the bound+stride may overflow.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23511 3a8e244f-b0f2-452b-bcba-4c88e055c3ca