Extensions to code for tracking the time spent within the compiler
1) Introduced a new class PhaseTracker which collects the time
stamp at selected points during the execution of the compiler.


The PhaseTracker collects timestamps for calls to either of

  void StartPhase(const char* phaseName);
  void StartPhase(const char* passName, SubPhase subPhase);

are made.  The former is used outside runpasses() and the latter
is used at several points during a single call to runpass().


If the user adds either --print-passes or --print-passes-file on
the command line the PhaseTracker will emit a message at
the end of each phase, and then a more detailed pair of reports
at the end of the run.






2) Inserted a single PhaseTracker immediately after main() in driver.cpp
and collect the results just before clean_exit()

Defined a new "phase" immediately before the driver starts the
initialization phase and immediately after the return from runpasses.




3) Pulled out the old, and slighly clumsy, code for tracking time in
runpasses/runpass and replaced it with calls to the PhaseTracker.



4) If --print-passes or --print-passes-file is added to the command line
then the PhaseTracker will print the same output as these flags do today
but will also generate two new reports:

   a) A report with the time for every subphase or every pass sorted by
   pass.  This report also shows the percentage of the total time spent
   in that pass and the running total of these percentages.

   b) A similar report to (a) but sorted by the duration of each pass 
   with the slowest passes reported first.



5) Moved the declarations for printPasses and printPassesFile from 
runpasses.{h,cpp} to driver.{h,cpp}




The goal of offering more detailed information surfaced last week during
friendly discussion regarding the portion of time spent by various phases.
The new reports make it easy to track this.

The goal of ensuring that we capture all time from the entry to main()
to the call to clean_exit() arose from a recent regression that appears
to have caused setupOrderedGlobals() to become the second slowest
"pass" for simple Chapel programs and that may also be adding overhead
to makeBinary.  The production print-passes highlights the additional
time taken in makeBinary but did not reveal the new overhead in during
initialization.  This wasn't full appreciated until we ran "time chpl ..."
and compared the wall time with the reported time.


Compiled:  STD with clang on Mac and STD/LLVM with gcc on chap01

Tested:    Multiple runs on several small programs
           Paratest
           Elliot tested against the new output when commiting his
           patch for the nightly performance reporting

Reviewed:  Elliot




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23519 3a8e244f-b0f2-452b-bcba-4c88e055c3ca