Remove some more timeout files
[ok'd by perf team]

This continues a general cleaning up of the timeout files in the testing
directory. The primary reason is to get tests that actually require the timeout
down to under the default timeout to decrease testing time.

All of these tests currently appear to finish well under their timeouts so the
timeouts are being removed. Some of these were previously performance tests so
they might have taken longer, but overall I think compilation time/execution
time has decreased substantially for some tests in the last year or so that the
timeouts aren't needed anymore.

There's ~5 tests that still have timeouts 4 of which have known reasons (and
another 7 that have timeouts smaller than the default, which we are going to
leave in.)

Removed files and associated timeouts are:
  2000 - release/examples/benchmarks/hpcc/variants/TIMEOUT
  2000 - studies/hpcc/CommDiags/TIMEOUT
  600  - release/examples/benchmarks/miniMD/PERFTIMEOUT
  600  - release/examples/benchmarks/miniMD/explicit/PERFTIMEOUT

It's possible they might time out under some configurations and in particular
may time out compilation with the cray compiler. In such a case I can look into
the cause and at the very least put back in a much smaller timeout with a note
indicating why it is there.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23528 3a8e244f-b0f2-452b-bcba-4c88e055c3ca