Correctly initialize all member variables; oversight highlighted by Coverity
I failed to initialize one of the member variables in an internal class
used by the new PhaseTracker.  Although this variable was in fact written
before it was read, it is good to assure that the instances are internally
consistent.

Compiled:  STD on clang
Tested:    Generated timeing information of Sequential Triad
Reviewed:  Elliot




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23529 3a8e244f-b0f2-452b-bcba-4c88e055c3ca