Update endCountTest/taskCount
The fork counts in these two tests (same source, taskCount uses
endCountTest) are racy by nature, so I changed the tests to be a
little less sensitive to false negatives in matching the .bad file and
also to print only the comm counts we care about to make maintaining
them a little easier.

For taskCount, we don't actually care about the number of forks, so
just print out the number of gets and puts, which is deterministic.
For endCountTest, we only print out the fork counts but also remove
the .bad file.

- changed endCountTest.chpl to enable selective printing of comm
  counts
- added compopts for both tests to enable the appropriate printing
- updated .good and .bad files as needed



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23530 3a8e244f-b0f2-452b-bcba-4c88e055c3ca