Factor out symbol replacement in callDestructor's changeRetToArgAndClone [reviewed by hilde]
I noticed that two parts of this if statement were identical beyond the
conditions that caused it to enter the statement, so I factored it out into a
helper function.  Additionally, Tom pointed out that the first branch was dead
code, as there should no longer be any instances of the (move lhs (= lhs rhs))
assignment left in the code, so after testing, that branch has been removed.

Hopefully, I can make use of this function when I add replacement for the
returned symbol in the callee.

tested on std/, and no-local examples/


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23533 3a8e244f-b0f2-452b-bcba-4c88e055c3ca