Trivial tweak to the formatting of output for --print-passes and --print-passes-file
The new output for --print-passes-file produces 3 tables of values where the old
format generated just one table.  Elliot and I shifted our opinions about the
format of these tables once or twice with the result that one row in the old-style
table changed its name.  This turned out to have much larger consequences for
Elliot than we initially anticipated.

This very minor change switches back the old column name.

Compiled: STD on clang (mac) and STD on gcc (chap01)

Tested:   Verified output on the report
          Ran paratest for about 15 minutes (**)



** I cannot conceive of a world, at least not on I that I want to live in,
in which this change could break the build.  Additionally I am preparing to
kick off a full run of paratest for 3 Coverity changes and so will be one
of the first victims if this proves to be wrong.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23537 3a8e244f-b0f2-452b-bcba-4c88e055c3ca