Look for 'total time: ' instead of 'total :' for compiler performance keys
[trivial, not reviewed] 

'total time' was changed to 'total' with r23519. I updated most of the compiler
performance .dat files to reflect this (and due to this and other changes
chap08 and chap12 were blown away.) However I forgot about the compiler
performance tests that exist in regular performance testing.  Just reverting
'total' back to 'total time' is easier than changing ~50 .dat files and
rerunning releaseOverRelease performance results. I'll update the .dat files
for chap01 again and blow away chap08 and chap12 again since they only have a
day's worth of data. 

This patch just reverts the perfkeys to look for 'total time', as well as the
prediff for filtering out the extra print-passes info and reverts the graph
keys for the 'total time' graphs. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23539 3a8e244f-b0f2-452b-bcba-4c88e055c3ca