Minor fixes in syncing perf graphs and saving .dat files
[trivial, not reviewed]

While I was looking into the errors syncing to sourceforge today I noticed that
we don't close the log file if the syncPerfGraphs script returns because a
SUCCESS file wasn't found. Also removed trailing whitespace from syncPerfGraphs
while I was in there. In nightly, the call to add svn ignore to non-dat
files didn't cd to the svnPerf directory first.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23540 3a8e244f-b0f2-452b-bcba-4c88e055c3ca