Update multiple configuration option for performance graphs
[reviewed by kyle]

Originally all configurations would be shown by default and if you selected
new graphs and redisplayed, the state of the configuration checkbox was not
honored. (If you unchecked qthreads, and selected a new suite, the qthreads
series would still be displayed on the graphs.)

This updates that so that the state of the checkboxes are honored for redraws
and so that by default only the top configuration is enabled by default. So for
us this means that fifo is shown by default only and qthreads is only shown if
you explicitly check the qthreads configuration box. The reason is that the
qthreads results are more for us right now, and probably confusing to people
outside.

This adds the ability to prevent a redraw when we set the colors after sorting
the series. This prevents an additional render and should speed up loading
times for the graphs some.

This also moves setting default annotations into the setupAnnToggle function
and comments out the call to this function. The reason is that calls to
setAnnotations force the graph to redraw and by moving this into the call and
removing the function call we save the redraw time since we don't actually
graph any annotations. Previously the code that made the annotation button
visible was commented out to do the same thing, but that way did not prevent
setting the annotations.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23542 3a8e244f-b0f2-452b-bcba-4c88e055c3ca