Fix graph generation for chap04
[reviewed by kyle]

With the switch to doing qthreads perf tests right after fifo we ran into two
problems. One was that the releaseOverRelease graphs got blown away when the
qthreads graphs were synced. To fix this the releaseOverRelease graphs are also
generated and synced with the qthreads results too. The data being spliced in
is still for the fifo results, it just runs with qthreads so that when we rsync
to sourceforge we don't delete the folder since it's not on the local machine
anymore.

The other problem  was that when fifo testing finished the graphs it synced
only contain the fifo and not fifo+qthreads graph. The reason for this was
originally so that we wouldn't get a bunch of errors the first day when we
didn't have any qthreads results. To fix this, we now look for both the default
and qthreads configurations for the fifo run. "default" is a special keyword
for the performance runs with multiple configurations to indicate that it's at
the base performance directory and that the perfkeys don't have the description
tacked on.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23543 3a8e244f-b0f2-452b-bcba-4c88e055c3ca