Re-render dygraph after sorting the keys
[trivial, not reviewed]

Yesterday I tried to make dygraph load faster by pushing a re-render until
after the series were sorted, colors were updated, and the check for when
series should be hidden based on configuration. However, we need to re-render
after sorting the keys because this updates other internal properties that are
needed for sorting colors. Without this compiler performance graphs don't (or
any expanded graph) don't have the dashed line set. This just adds a re-render
(setting annotations to themselves is a no-op that causes a redraw) back after
sorting the series.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca