launchcmd-for-aprun-launcher script that will allow aprun launcher to reserve nodes per test.
This python script runs a test as a pbs batch job. Once the job is queued, it
waits for the output file to exist, prints it to stdout and then exits.

Includes update to sub_test to support launch commands with arguments
(i.e. -launchcmd ".../mylauncher --with-an-arg 123").

This has been running cray internal tests for several months now, but there is
nothing special or internal about it.

r: eronaghan

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23546 3a8e244f-b0f2-452b-bcba-4c88e055c3ca