Address a Coverity exception and do a little cleanup in preparation for	two more
1) passes/codegen.cpp

Coverity was confused by an #ifdef inside an if() and concluded that a variable
might be read without being written.  Shifted the variable initialization.




The remaining changes are some simple cleanup that I noted while starting work
on two Coverity exceptions in files.cpp.  I will be submitting the fixes for the
Coverity issues separately.


2) mysystem.{h,cpp}
   passes/doc.cpp
   util/files.cpp

The function mysystem takes an optional	paramater that is declared to be an
integer but is treated as if a bool.  Converted the parameter to be a bool.
Updated	the callers that were passing 0	or 1.


3) utils/files.cpp

I am going to fix two Coverity exceptions in here.  Before doing this I

  (a) cleaned up the order of #includes

  (b) removed the wrapper genIntFilename() and updated the call	sites to
      call genIntermediateFilename directly




Compiled:  STD on clang/mac  STD/LLVM on gcc/chap01
Tested:	   Ran paratest
Reviewed:  Tom





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23552 3a8e244f-b0f2-452b-bcba-4c88e055c3ca