Update launchcmd-for-aprun-launcher script to tolerate moab/torque fast dequeues.
We occasionally see tests that fail because the launchcmd script cannot find
the original job in qstat. We think this might be due to qstat/qsub dequeueing
the job very quickly after completion. So, if we lose track of the job, but we
end up with an output file, assume it completed successfully.

r: eronaghan

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23554 3a8e244f-b0f2-452b-bcba-4c88e055c3ca