Snapshot of the AST debugging code that provides a "node" oriented view of the AST
Add the preliminary code for displaying the AST in a node-oriented fashion.
This submit does not include the changes to driver.cpp needed to add flags
to enable this mode; this code is currently unreachable.


Compiled:  STD on clang, STD/LLVM on gcc
Tested:    Work incomplete but no obvious bugs on simple test programs
Reviewed:  None




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23563 3a8e244f-b0f2-452b-bcba-4c88e055c3ca