Change problem size for ptrans performance runs
[ok'd by perf team]

ptrans has been timing out on chap04 for long time and is pretty close to
timing out on the other machines. This sets the problem size to be numrows=5733
for performance runs.

Previously the values were based on amount of memory on the system set to:
 chap04:    ~28,000 rows
 chap03:    ~8,000 rows
 brac-lnx:   5,733 rows

5,733 rows is ~.5GB of memory. Such an odd number was chosen because it was
previously what was running on bradc-lnx and it will be nice to keep the
historical data for it.

.graph and perkeys were updated to put the values into a new .dat file since
the old performance data will be off. I'll svn copy the data from bradc-lnx to
the new .dat file too.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23570 3a8e244f-b0f2-452b-bcba-4c88e055c3ca