Remove "final" timeout file: release/examples/benchmarks/hpcc/TIMEOUT
[ok'd by perf team]

With r23570 ptrans will no longer take over 30 minutes on chap04 so we can get
rid of this timeout. I tested some of the others in the directory and I'm
pretty sure they don't need the timeout either, but we'll see how testing goes.
This is the last of the timeouts that will be removed in the near term.

There are still 11 left overall. 7 of those are staying because they have
timeouts smaller than the default.The other 4 are:

  600  - release/examples/benchmarks/shootout/threadring.perftimeout
  400  - studies/shootout/thread-ring/lydia/TIMEOUT
   - Should be able to remove with switch to qthreads

  500  - release/examples/benchmarks/ssca2/SSCA_main.timeout
  500  - distributions/dm/t5a.timeout
   - These are because the cray compiler is really slow. Should be able to
     remove when our compilation times get better.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23571 3a8e244f-b0f2-452b-bcba-4c88e055c3ca