Remove mbp-kbrady crontab entry, it has been migrated to new system.
The mbp-kbrady test scripts were also updated to use the current working directory,
instead of ~/chapel, and remove the stdout/stderr pipes to log file. The new
system guarantees the current dir is a clean working copy of trunk, and it
captures stdout/stderr.

These tests have been disabled, and will continue to be disabled in the current
system. The configuration has moved, though, so they will be available whenever
mac testing stabilizes.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23583 3a8e244f-b0f2-452b-bcba-4c88e055c3ca