Remove chap01 crontab entry, it has been migrated to new system.
The chap01 test scripts were also updated to use the current working directory,
instead of ~/chapel, and remove the stdout/stderr pipes to log file. The new
system guarantees the current dir is a clean working copy of trunk, and it
captures stdout/stderr.

This also removes the scripts to synchronize the crontabs across the testing
hosts. All of the jobs have been migrated to the new system. w00t!

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23586 3a8e244f-b0f2-452b-bcba-4c88e055c3ca