Rename type variables in resolution of _cast [no objections]
Using dst and src as the names for the first and second argument of the cast
did not seem intuitive to me.  Did dst mean you were writing there, or did it
mean that was the type destination specified?  oldType and newType give a
clearer indication of what those sides mean.

tested on std/


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23587 3a8e244f-b0f2-452b-bcba-4c88e055c3ca