Go back to removing .exec.out.tmp on successful perf tests
[reviewed by lydia]

In r22299, David made sub_test remove .exec.out.tmp files on
successful perf key matches; but then in r22591, I reverted this
because, given the lack of correctness testing/screening on
performance tests, it became difficult to check whether or not a
performance test really got the right output.  In that commit message
I re-expressed a desire to have a way to validate performance test
output.

Kyle then implemented such a validation in r22825, so now it seems
smarter to remove .exec.out.tmp files when everything matches rather
than to leave them around; especially since some of them are quite
large.

I was thinking that this would be more complicated than reverting the
previous patch since it relies on both the perfkeys match and the
validation match.  But since both of those things are done by
computePerStats, it turned out to simply involve re-inserting David's
unlink call.  Nice.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23595 3a8e244f-b0f2-452b-bcba-4c88e055c3ca