Minor mod to avoid warnings if CHPL_HOME not set
[reviewed by Kyle]

Sung pointed out today that we're getting warnings now when
building with CHPL_HOME unset, which is how we suggest users
build.  This is a simple fix to make that go away.  Kyle's
larger rework of the chplenv/ scripts from Perl to Python
will also fix this, but that's still a few days away.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23603 3a8e244f-b0f2-452b-bcba-4c88e055c3ca