Swapping order of performance and release testing on bradc-lnx
Performance testing has been growing slightly in time recently, so
let's run it before release testing to minimize the chances that it's
still coming when I come in and start disturbing my machine.

Longer-term, we've discussed having the normal correctness testing run
from the release tarball and mixing other aspects of the test release
script into a smoketest after the creation of the release tarball.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23607 3a8e244f-b0f2-452b-bcba-4c88e055c3ca