Remove dead flags FLAG_DATA_SET_ERROR, FLAG_FIXED_STRING and FLAG_HEAP_ALLOCATE.
[Reviewed and committing for Tom]

FLAG_HEAP_ALLOCATE was added in r14701 by deitz, and removed in r20807
in favor FLAG_COFORALL_INDEX_VAR but left around in case it could be
reused.

FLAG_FIXED_STRING was added in r17296 by jonathanturner.  It is used only in
test code as a pragma for an experimental fixed-length string implementation.
One test (fixedStringEquals) is a future, and the other
(missingCastThisPragmaError) just prints an error message.  It does not appear
to be a complete implementation.

FLAG_DATA_SET_ERROR has been in flags.h since the beginning of time (r14559),
and is currently unused in the compiler, modules or test code.

[tested on / (--no-local): No failures.]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23615 3a8e244f-b0f2-452b-bcba-4c88e055c3ca