Changed comment "style" of a multi-line comment
[not reviewed, but discussed with Lydia]

This comment started with a /** and ended with a */, which turns out
to tickle a bug in chpldocs comment parsing.  Changed the /** to /*.

This comment was actually moved verbatim from ChapelBase.chpl in
r23558.  The reason this bug never showed up before is because we
don't parse the internal modules in non-developer mode.

FYI, Lydia will be filing a future for this bug in a separate commit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23618 3a8e244f-b0f2-452b-bcba-4c88e055c3ca