Adding a future for chpldoc specified comments.
Sung exposed a bug by having a comment which started with the specified header
of chpldoc/compflags/comment/basic.chpl and loseComments.chpl but did not close
in the same style.  This led to a valgrind error because it was assumed that
these comments would match.  We believe the proper behavior is to drop such
comments when discovered, but warn the user that we are doing so.  However, this
is a low priority bug at the moment.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23624 3a8e244f-b0f2-452b-bcba-4c88e055c3ca