Minor revision to main/arg.cpp
Separated the processing of environment variables into two functions

 a) The first function scans the descriptors and finds any environment
    variables and, in one case, "parses" the environment value in a flag
    sensitive manner.

 b) The second function applies the resulting value
    i)  To the "location" if the location is not NULL
    ii) Invokes the handler function if present.


This second portion can be re-used when processing the CommandLine.


Also spotted a bug in one entry of the descriptors in driver; a "location"
pointer was set up in the descriptor "as if" a char* but was actually a
FILE*. The descriptor "type" was set in a way that caused this discrepancy
to be overlooked.



Compiled: STD clang    STD/LLVM gcc

Tested:   Added some printfs to the code and tested an example of every
          "type" of environment variable.

          Ran paratest


Reviewed: None.  Very simple partitioning.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23629 3a8e244f-b0f2-452b-bcba-4c88e055c3ca