Fix for Coverity bug 1212390 [simple, not reviewed]
Coverity complained that in the method to rename the instantiated type, space
was allocated in a character buffer for the size the name would be after the
change, but that the character sequence that was written was longer than that
(due to the null terminator and because this rename removes the last character).
To fix, I caused the size of memory allocated to be one character greater, thus
taking the old null terminator's position into account.

tested over std/



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23635 3a8e244f-b0f2-452b-bcba-4c88e055c3ca