Rework printchplenv and util/chplenv/*
[reviewed by elliot]

This patch converts all of the util/chplenv/* scripts into python. The
motivation for this change was to reduce the overhead in our Makefiles
caused by popens to other scripts. Now that all of the scripts are in
python, they just import each other and call a method. The end result is
cutting ~1.5s off of every compile.

To do this I had to rename all of the chplenv/* scripts. So, if you call any
of them from your .bashrc or other personal scripts, they will need to be
updated.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23638 3a8e244f-b0f2-452b-bcba-4c88e055c3ca