Adding a multilocale c_string futures
There are 2 types of futures here.

The first is an explicit test for an existing gasnet regression
(cannot create a C string from a remote string).  In this case, the
Chapel string should be copied to the remote locale before calling
c_str().

The other is a case where the compiler generates chpl_gen_comm_put()
calls to copy c_strings variables, which is wrong in 2 ways.  The
string is put in a remote c_string variable which has no space
allocated for it, and it only sends sizeof(c_string) bytes of the
string.  The long term solution is to mark all c_string variable as
"local" in the compiler (something we want in the language and
compiler for performance reasons), and then generate a runtime error
when we try to access it from a locale other than the one it is
allocated on.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23639 3a8e244f-b0f2-452b-bcba-4c88e055c3ca