Consolidate system() calls in driver.cpp
[reviewed by mike]

This patch consolidates the 23 system() calls used to set up the env vars
in driver.cpp into a single popen out to printchplenv. Each of those calls
also created a temporary file that was used to read the output back into
the compiler. In my tests on the chap boxes, the time in init dropped from
~.5s -> ~.05s after this change, the difference is even larger on my mac.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23649 3a8e244f-b0f2-452b-bcba-4c88e055c3ca