Stop appending performance description to perfkeys
[trivial, not reviewed]

For the multi configuration performance testing I was originally adding
the performance description to all of the perf keys (so if you had <key>
previously you would have <key>.perfkey now.) This only really affected the
perfkeys in the .dat files. I originally thought the perfkeys had to be unique,
but that's only true for keys that are coming from the same file. Since
descriptions for multiple configurations come from different files we don't
have to do this.

This just reverts my changes to computePerfStats from r23502 and the code in
sub_test that passed arguements to coputePerfStats.

I'll change the headers in the .dat files to reflect the lack of descriptions
after perfkeys.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23658 3a8e244f-b0f2-452b-bcba-4c88e055c3ca